Changed from rep/req sockets to dealer/router sockets for download and upload of files.

+++ Tarea3/fileClient.cc
@@ -79,14 +79,14 @@ void getCheckSum(string path, unsigned char *check_sum) {
 
 void CheckFile(socket &download_socket) {
 	message check_sum_msg;
-	string path, fname;
+	string path, fname, server_identity, empty;
 	unsigned char check_sum[SHA_DIGEST_LENGTH];
 	unsigned char *received_check_sum = NULL;
 
 	download_socket.receive(check_sum_msg);
 
-	check_sum_msg >> fname;
-	received_check_sum = (unsigned char *)check_sum_msg.raw_data(1);
+	check_sum_msg >> empty >> fname;
+	received_check_sum = (unsigned char *)check_sum_msg.raw_data(2);
 	cout << "Received Check sum: ";
 	printChecksum(received_check_sum);
 
@@ -120,21 +120,20 @@ void CheckFile(socket &download_socket) {
 }
 
 int SaveFile(socket &download_socket) {
-	//cout << "total: " << total_downloaded << "; chunks: " << chunks_downloaded << "; file size: " << file_size << "; downloading: " << DOWNLOADING << endl;
 	message answer, req_check_sum;
-	string path, fname;
+	string path, fname, server_identity, empty;
 	size_t size;
 	int size_chunk;
 	char *data;
 	FILE* f;
 	double progress;
 
-	cout << "Saving to file..." << endl;
-
 	download_socket.receive(answer);
 
-	if(answer.get(0) == "Error") {
-		if(atoi(answer.get(1).c_str()) == 0) {
+	answer >> empty;
+
+	if(answer.get(1) == "Error") {
+		if(atoi(answer.get(2).c_str()) == 0) {
 			cout << "\nThe file requested does not exist or couldn't be read!!" << endl;
 			return -1;
 		}
@@ -162,14 +161,14 @@ int SaveFile(socket &download_socket) {
 		f = fopen(path.c_str(), "wb");
 		answer >> file_size;
 		answer >> size_chunk;
-		size = answer.size(3);
-		data = (char*) answer.raw_data(3);
+		size = answer.size(4);
+		data = (char*) answer.raw_data(4);
 	} else {
 		f = fopen(path.c_str(), "ab");
 		fseek(f, 0L, SEEK_END);
 		answer >> size_chunk;
-		size = answer.size(2);
-		data = (char*) answer.raw_data(2);
+		size = answer.size(3);
+		data = (char*) answer.raw_data(3);
 	}
 
 	fwrite(data, 1, size, f);
@@ -192,179 +191,75 @@ int SaveFile(socket &download_socket) {
 }
 
 void downloadFile(socket &broker_socket, socket &download_socket, string op, string username) {
-	if(DOWNLOADING) {
-		cout << "you are already downloading a file, wait untill that ends to request another download!\n";
-		return;
-	}
-
-	//cout << "total: " << total_downloaded << "; chunks: " << chunks_downloaded << "; file size: " << file_size << "; downloading: " << DOWNLOADING << endl;
-
 	char fname[100];
 	string server_address;
 	string path;
 	message send_request, broker_response, metadata_file, req_check_sum;
 
+	if(DOWNLOADING) {
+		cout << "you are already downloading a file, wait untill that ends to request another download!\n";
+		return;
+	}
+
 	cout << "Enter file name: \n";
 	cin.getline(fname, sizeof(fname));
 	cin.getline(fname, sizeof(fname));
+	//getline(cin, fname);
 
-	cout << op << " " << username << " " << fname << endl; //send req to broker
-	//cin >> server_address; //rec response from broker
-	//download_socket.connect(server_address); //connect to server
+	cout << "enter server address: " << endl; //send req to broker
+	cin >> server_address; //rec response from broker
+	download_socket.connect(server_address); //connect to server
 	DOWNLOADING = true;
-
-	metadata_file << op << username << fname << total_downloaded;
-	while(true) {
-		broker_socket.send(metadata_file);
-		int stat = SaveFile(broker_socket);
-		if(stat == 1) {
-			metadata_file << op << username << fname << total_downloaded;
-		} else if(stat == 0 || stat == -1) {
-			break;
-		}
-	}
-
-	if((int)total_downloaded == file_size) {
-		req_check_sum << op << username << fname << total_downloaded;
-		broker_socket.send(req_check_sum);
-		CheckFile(broker_socket);
-		cout << "total: " << total_downloaded << "; chunks: " << chunks_downloaded << "; file size: " << file_size << "; downloading: " << DOWNLOADING << endl;
-		total_downloaded = 0;
-		chunks_downloaded = 0;
-		file_size=0;
-		DOWNLOADING = false;
-	}
-
-	/*end of download*/
-
-
-	/*message answer;
-	size_t size;
-	FILE* f;
-	int size_chunk;
-	long file_size;
-	double progress;
-	string tmp;
-	char *data;
-
-	size_t total = 0;
-	size_t chunks = 0;
-
-	unsigned char check_sum[SHA_DIGEST_LENGTH];
-	unsigned char *received_check_sum = NULL;
-
-	//Check if "Downloads/" dir exists,
-	//	if not, it's created.
-	struct passwd *pw = getpwuid(getuid());
-	const char *homedir = pw->pw_dir;
-
-	path = (string)homedir + "/Descargas/Downloads/";
-
-	struct stat sb;
-	lstat(path.c_str(), &sb);
-
-	if(!S_ISDIR(sb.st_mode)) {
-		string url = "mkdir -p " + (string)path;
-		const char * directory =  url.c_str();
-		system(directory);
-	}
-
-	path.append(fname);
-
+	bool error = false;
 	cout << "Saving to file..." << endl;
 
-	SHA_CTX sha_ctx;
-	SHA1_Init(&sha_ctx);
-
+	metadata_file << "" << op << username << fname << total_downloaded;
 	while(true) {
-		metadata_file << op << username << fname << total;
-		broker_socket.send(metadata_file);
-
-		broker_socket.receive(answer);
-
-		if(answer.get(0) == "Error") {
-			if(atoi(answer.get(1).c_str()) == 0) {
-				cout << "\nThe file requested does not exist or couldn't be read!!" << endl;
-				return;
-			}
-		}
-
-		answer >> tmp;
-
-		if (total == 0) {
-			f = fopen(path.c_str(), "wb");
-			answer >> file_size;
-			answer >> size_chunk;
-			size = answer.size(3);
-			data = (char*) answer.raw_data(3);
-		} else {
-			f = fopen(path.c_str(), "ab");
-			fseek(f, 0L, SEEK_END);
-			answer >> size_chunk;
-			size = answer.size(2);
-			data = (char*) answer.raw_data(2);
-		}
-
-		fflush(f);
-
-		chunks++;
-		fwrite(data, 1, size, f);
-		SHA1_Update(&sha_ctx, data, size);
-		total += size;
-		fclose(f);
-
-		progress = ((double)total/(double)file_size);
-		if((progress*100.0) <= 100.0){
-			printf("\r[%3.2f%%]",progress*100.0);
-		}
-		fflush(stdout);
-
-		if (size == 0 || (int)size < size_chunk) {
-			req_check_sum << op << username << fname << total;
-			broker_socket.send(req_check_sum);
+		download_socket.send(metadata_file);
+		int stat = SaveFile(download_socket);
+		if(stat == 1) {
+			metadata_file << "" << op << username << fname << total_downloaded;
+		} else if(stat == 0) {
+			break;
+		} else if(stat == -1) {
+			error = true;
 			break;
 		}
 	}
 
-	SHA1_Final(check_sum, &sha_ctx);
+	if((int)total_downloaded == file_size && !error) {
+		req_check_sum << "" << op << username << fname << total_downloaded;
+		download_socket.send(req_check_sum);
+		CheckFile(download_socket);
+		//cout << "total: " << total_downloaded << "; chunks: " << chunks_downloaded << "; file size: " << file_size << "; downloading: " << DOWNLOADING << endl;
+	}
 
-	cout << endl;
-	cout << "Chunks received: " << chunks << "\nBytes received: " << total << endl;
-
-	broker_socket.receive(answer);
-	answer >> tmp;
-	received_check_sum = (unsigned char *)answer.raw_data(1);
-
-	cout << "Received Check sum: ";
-	printChecksum(received_check_sum);
-
-	cout << "Calculated check sum: ";
-	printChecksum(check_sum);
-
-	if(memcmp(check_sum, received_check_sum, SHA_DIGEST_LENGTH)) {
-		cout << "\nThe Checksum failed, please try again!!" << endl;
-		if( remove(path.c_str()) != 0 )
-			cout << "Error deleting file: " << path << endl;
-		else
-			cout << "File: " << path << " successfully deleted\n";
-	} else {
-		cout << "File saved successfully!!\n";
-	}*/
+	total_downloaded = 0;
+	chunks_downloaded = 0;
+	file_size=0;
+	DOWNLOADING = false;
+	download_socket.disconnect(server_address);
 }
 
 void uploadFile(socket &broker_socket, socket &download_socket, string op, string username) {
 	FILE* f;
 	char *data;
 	size_t size, offset;
-	char fname[100], file_name[100];
+	char file_name[100], fname[100];
 	message file_message, server_answer, check_sum_msg;
 	double progress;
+	string empty, server_address;
 
 	unsigned char check_sum[SHA_DIGEST_LENGTH];
+
+	cout << "enter server address: " << endl; //send req to broker
+	cin >> server_address; //rec response from broker
+	download_socket.connect(server_address); //connect to server
 
 	cout << "Enter file name: \n";
 	cin.getline(fname, sizeof(fname));
 	cin.getline(fname, sizeof(fname));
+	//cin.ignore(numeric_limits<streamsize>::max(), '\n');
 
 	getFileName(fname, file_name);
 
@@ -392,14 +287,16 @@ void uploadFile(socket &broker_socket, socket &download_socket, string op, strin
 		size = fread(data, 1, CHUNK_SIZE, f);
 		SHA1_Update(&sha_ctx, data, size);
 
-		file_message << op << username << file_name << CHUNK_SIZE << offset;
+		file_message << "" << op << username << file_name << CHUNK_SIZE << offset;
 
 		file_message.push_back(data, size);
 		free(data);
 
-		broker_socket.send(file_message);
+		download_socket.send(file_message);
 
-		broker_socket.receive(server_answer);
+		download_socket.receive(server_answer);
+
+		server_answer >> empty;
 
 		offset += size;
 		progress = ((double)offset/(double)sz);
@@ -408,12 +305,12 @@ void uploadFile(socket &broker_socket, socket &download_socket, string op, strin
 		}
 		fflush(stdout);
 
-		if(server_answer.get(0) == "Error") {
-			if(atoi(server_answer.get(1).c_str()) == 0) {
+		if(server_answer.get(1) == "Error") {
+			if(atoi(server_answer.get(2).c_str()) == 0) {
 			cout << "\nThe file requested does not exist or couldn't be read!!" << endl;
 			return;
 			}
-		} else if(server_answer.get(0) == "Done") {
+		} else if(server_answer.get(1) == "Done") {
 			fclose(f);
 			break;
 		}
@@ -424,24 +321,26 @@ void uploadFile(socket &broker_socket, socket &download_socket, string op, strin
     cout << "\nCalculated check sum: ";
 	printChecksum(check_sum);
 
-	check_sum_msg << op << username << file_name << "Check";
+	check_sum_msg << "" << op << username << file_name << "Check";
 
     check_sum_msg.push_back(check_sum, SHA_DIGEST_LENGTH);
 
-    broker_socket.send(check_sum_msg);
+    download_socket.send(check_sum_msg);
 
-	broker_socket.receive(server_answer);
+	download_socket.receive(server_answer);
 
-	if(server_answer.get(0) == "Error") {
-		if(atoi(server_answer.get(1).c_str()) == 0) {
+	server_answer >> empty;
+
+	if(server_answer.get(1) == "Error") {
+		if(atoi(server_answer.get(2).c_str()) == 0) {
 			cout << "\nThe file requested does not exist or couldn't be read!!" << endl;
 			return;
-		} else if(atoi(server_answer.get(1).c_str()) == 0) {
+		} else if(atoi(server_answer.get(2).c_str()) == 1) {
 			cout << "\nThe Checksum failed, please try again!!" << endl;
 		}
 	}
 
-	if(server_answer.get(0) == "ok") {
+	if(server_answer.get(1) == "ok") {
 		cout << "File has been uploaded successfully!!" << endl;
 	}
 }
@@ -488,10 +387,17 @@ int main() {
 
 	context ctx;
 	socket broker_socket(ctx, socket_type::req);
-	socket download_socket(ctx, socket_type::pull);
+	socket s(ctx, socket_type::dealer);
 
-	socket upload_socket(ctx, socket_type::push);
-	upload_socket.bind("tcp://*:5556");
+	zmq_pollitem_t standardin;
+	standardin.socket = NULL;
+	standardin.fd = 0;
+	standardin.events = poller::poll_in;
+
+	poller p;
+
+	p.add(broker_socket, poller::poll_in);
+	p.add(standardin);
 
 	cout << "Connecting to tcp port 5555\n";
 	broker_socket.connect("tcp://localhost:5555");
@@ -542,28 +448,37 @@ int main() {
 			printMenu();
 			cout <<  "Enter action to perform: \n";
 			cin >> op;
-
-			if(op == "Upload" || op == "upload" || op == "up") {
-				uploadFile(broker_socket, upload_socket, "Upload", username);
-			} else if(op == "Download" || op == "download" || op == "down") {
-				downloadFile(broker_socket, download_socket, "Download", username);
-			} else if(op == "List_files" || op == "list_files" || op == "ls") {
-				listFiles(broker_socket, "List_files", username);
-			} else if(op == "Delete" || op == "delete" || op == "del") {
-				deleteFile(broker_socket, "Delete", username);
-			} else if(op == "Exit" || op == "exit" || op == "ex") {
-				break;
-			} else {
-				cout << "\nInvalid option, please enter one of the listed options!\n";
-			}
+			/*if(p.poll()) {
+				if(p.has_input(broker_socket)) {
+					//do something eith message
+					cout <<"Message" << endl;
+				}
+				if(p.has_input(standardin)) {
+					getline(cin, op);*/
+					if(op == "Upload" || op == "upload" || op == "up") {
+						uploadFile(broker_socket, s, "Upload", username);
+						p.add(s, poller::poll_out);
+					} else if(op == "Download" || op == "download" || op == "down") {
+						downloadFile(broker_socket, s, "Download", username);
+						p.add(s, poller::poll_in);
+					} else if(op == "List_files" || op == "list_files" || op == "ls") {
+						listFiles(broker_socket, "List_files", username);
+					} else if(op == "Delete" || op == "delete" || op == "del") {
+						deleteFile(broker_socket, "Delete", username);
+					} else if(op == "Exit" || op == "exit" || op == "ex") {
+						break;
+					} else {
+						cout << "\nInvalid option, please enter one of the listed options!\n";
+					}
+			//	}
+			//}
 		}
 	}
 	else cout << "\nUser not found!!\n\n";
 
 	cout << "Finished\n";
 	broker_socket.close();
-	download_socket.close();
-	upload_socket.close();
+	s.close();
 	ctx.terminate();
 	return 0;
 }
